+++ src/vqueue/src/main.rs
@@ -5,6 +5,16 @@ use vsmtp_config::Config;
 fn main() -> anyhow::Result<()> {
     let args = <Args as clap::StructOpt>::parse();
 
+    if args.version {
+        println!(
+            "{} v{}\ncommit: {}",
+            clap::crate_name!(),
+            clap::crate_version!(),
+            env!("GIT_HASH")
+        );
+        return Ok(());
+    }
+
     let config = args.config.as_ref().map_or_else(
         || Ok(Config::default()),
         |config| {
@@ -15,5 +25,9 @@ fn main() -> anyhow::Result<()> {
         },
     )?;
 
-    execute(args.command, &config)
+    if let Some(command) = args.command {
+        execute(command, &config)
+    } else {
+        anyhow::bail!("no commands where specified")
+    }
 }